unify diff/rounding

+++ api/src/main/java/at/outdated/bitcoin/exchange/api/performance/AbsoluteComparator.java
@@ -13,6 +13,6 @@ public class AbsoluteComparator implements Comparator<Performance> {
 
     @Override
     public int compare(Performance o1, Performance o2) {
-        return (int) Math.round( (o1.getTotalDifference()- o2.getTotalDifference())*100.0 );
+        return Math.round( (float)(o1.getTotalDifference()- o2.getTotalDifference())*100.0f );
     }
 }
